refactor: move core functions in core crate
