--- cli/src/commands/list/mod.rs
@@ -1,3 +1,0 @@
-mod list_command;
-
-pub use list_command::*;
