+++ cli/src/commands/list_command.rs
@@ -1,7 +1,9 @@
-use std::{path::PathBuf, time::Instant};
+use std::time::Instant;
 
-use acevo_content_editor::models::{Benchmarked, FileEntry, PackageFileTable};
-use glob::Pattern;
+use acevo_content_editor::{
+	functions::list_query,
+	models::{Benchmarked, FileEntry},
+};
 
 use crate::args::{GlobalOpts, ListArgs};
 
@@ -26,10 +28,4 @@ impl ListCommand {
 			data: entries,
 		}
 	}
-}
-
-pub fn list_query(path: impl Into<PathBuf>, pattern: &Pattern) -> Vec<FileEntry> {
-	let path = path.into();
-	let file_table = PackageFileTable::read_unpacked_from(path);
-	file_table.query(pattern)
 }
