+++ cli/src/commands/mod.rs
@@ -1,5 +1,5 @@
-mod list;
-mod unpack;
+mod list_command;
+mod unpack_command;
 
-pub use list::*;
-pub use unpack::*;
+pub use list_command::*;
+pub use unpack_command::*;
