--- cli/src/commands/unpack/mod.rs
@@ -1,7 +1,0 @@
-mod unpack_command;
-mod unpack_task_data;
-mod unpack_task_response;
-
-pub use unpack_command::*;
-pub use unpack_task_data::*;
-pub use unpack_task_response::*;
