+++ core/src/models/mod.rs
@@ -1,11 +1,17 @@
+mod benchmarked;
 mod file_entry;
 mod package_file_table;
 mod packed_package_buffer;
 mod path_tree;
-mod benchmarked;
+#[cfg(feature = "async")]
+mod unpack_task_data;
+mod unpack_task_response;
 
+pub use benchmarked::*;
 pub use file_entry::*;
 pub use package_file_table::*;
 pub use packed_package_buffer::*;
 pub use path_tree::*;
-pub use benchmarked::*;
+#[cfg(feature = "async")]
+pub use unpack_task_data::*;
+pub use unpack_task_response::*;
