+++ core/src/models/unpack_task_data.rs
@@ -1,7 +1,8 @@
 use std::{path::PathBuf, sync::Arc};
 
-use acevo_content_editor::models::FileEntry;
 use tokio::{fs::File, sync::Mutex};
+
+use super::FileEntry;
 
 pub struct UnpackTaskData {
 	pub entry: FileEntry,
